start refactoring kafka
