+++ backend/bin/kafkaConsumer/userCourseProgressConsumer/saveToDB.ts
@@ -6,10 +6,10 @@ import {
   UserCourseProgress,
   UserCourseServiceProgress,
 } from "@prisma/client"
-import TmcClient from "../../../services/tmc"
 import { generateUserCourseProgress } from "./generateUserCourseProgress"
 import { Logger } from "winston"
 import { pushMessageToClient, MessageType } from "../../../wsServer"
+import getUserFromTMC from "../../common/getUserFromTMC"
 
 import _KnexConstructor from "knex"
 
@@ -27,22 +27,6 @@ const Knex = _KnexConstructor({
       ? ["moocfi$production"]
       : ["default$default"],
 })
-
-const getUserFromTMC = async (prisma: PrismaClient, user_id: number) => {
-  const tmc: TmcClient = new TmcClient()
-  const userDetails = await tmc.getUserDetailsById(user_id)
-
-  return prisma.user.create({
-    data: {
-      upstream_id: userDetails.id,
-      first_name: userDetails.user_field.first_name,
-      last_name: userDetails.user_field.last_name,
-      email: userDetails.email,
-      username: userDetails.username,
-      administrator: userDetails.administrator,
-    },
-  })
-}
 
 export const saveToDatabase = async (
   message: Message,
